Add: about, xr-business, new 에 활용할 컴포넌트 파일 구현
