+++ src/App.jsx
@@ -4,7 +4,7 @@ import Header from './components/Header/Header';
 import Home from './pages/Home';
 import Login from './pages/Auth/Login';
 import Signup from './pages/Auth/Signup';
-import About from './pages/About';
+import About from './pages/Editer/About';
 
 function App() {
   return (
